fix(swingset): split two files for upcoming refactoring
`test-weakcollections.js` was split into one file that needs a full
kernel, and a separate file which only needs liveslots.

`tools/vo-test-harness.js` was split similarly.

The liveslots-only portions of both will be moved to a new package
soon, and this simplifies the diff.

Other packages import `tools/vo-test-harness.js`, but they use only
the remaining (full-kernel) portion. Nothing external uses the
liveslots-only portion.
